Fixes #8695

+++ lib/functions.zsh
@@ -8,7 +8,9 @@ function uninstall_oh_my_zsh() {
 
 function upgrade_oh_my_zsh() {
   env ZSH="$ZSH" sh "$ZSH/tools/upgrade.sh"
-  rm -rf "$ZSH/log/update.lock"
+  if [ -f "$ZSH/log/update.lock" ]; then
+    rm -rf "$ZSH/log/update.lock"
+  fi
 }
 
 function take() {
